api_service/requirements.txt
# Основной веб-фреймворк
fastapi==0.104.1

# ASGI-сервер для запуска FastAPI (без uvloop для совместимости с Windows)
uvicorn==0.24.0

# Драйверы для баз данных
elasticsearch[async]==8.10.0
redis==5.0.1

# Для работы с .env файлами
python-dotenv==1.0.0

# Все остальные пакеты (pydantic, orjson, starlette и т.д.)
# будут установлены автоматически как зависимости указанных выше.

# Для повторных попыток подключения
backoff==2.2.1